fix: move CCSharing enum to types.ts
